New File Dialog: Tweak default node on Projct Page
The case from the bug report is:
/
 --> project.pro
 --> src
     --> foo.qrc
     --> folder
         --> file added here

Since the .qrc file is nearer then the .pro file, it was selected as the default node.
This patch makes the priority for the .qrc node -1 if it's a unmatched mime type.
That means a .qrc node will only be the default if it's either a image, a .qml or a .js
file.

Task-number: QTCREATORBUG-12793
Change-Id: I7328fd6359cc2e1959ab0bda831bd77de117baff
Reviewed-by: Eike Ziller <eike.ziller@digia.com>