QmlProfiler: Rename JavaScript view to "V8"
The fact that the separate JavaScript view only works with versions of
Qt that use the V8 JavaScript engine is frequently confusing users.
JavaScript events for newer versions of Qt are integrated into the
Events and Timeline views, making the separate JavaScript view obsolete.
For those users who already know it and who are still using Qt 5.0 and
5.1 "V8" will be recognizable. For others "V8" won't mean anything and
they'll ignore it.

Change-Id: Iac0d89db24a3677e47b978a152cfc25c01f9c2b6
Task-number: QTCREATORBUG-12766
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@digia.com>
Reviewed-by: Robert Loehning <robert.loehning@digia.com>