QmlDesigner.PropertyEditor: Commit color string if selection changes
The LineEdit in the ColorEditor is not binded to a backendvalue
and we react manually to isAccepted (writeValueManually is set to true).

Because of this (and the famous focus bugs), we do not set the color,
if the selection is changed or the user enters edit mode without
pressing enter.

This patch adds the signal onCommitData to LineEdit and we react to
the signal. Because the the change of the backendvalue is delayed by
a timer we set the color immediately on the backend if we are not in gradient
editing mode.

Task-number: QTCREATORBUG-12841
Change-Id: Ic98640398c2c46b5e529780b4a9638c90760063f
Reviewed-by: Tim Jenssen <tim.jenssen@digia.com>