Utils: bring QtConcurrent::run into the current century.
.. by adding support to pass in a std::function (which includes lambdas)
as the function to run. The run function only takes functors that have
a QFutureInterface& as parameter, because a lambda can already capture
anything that would be passed in on invocation. Parameter-less
std::function<>s are already handled by QtConcurrent (in Qt).

To use it, you'll have to explicitly set the template parameter T on the
call.

Change-Id: Ie145fc3245d59fdea9d4703dab155ce9100eff89
Reviewed-by: Daniel Teske <daniel.teske@digia.com>