QmlDesigner: Fix regression for import paths
The imports paths used by the sub component manger and meta info system
were not correctly resolved. The main reason was that the textDocument in
the text modifiers are just plain text buffers and do not containt
any url for the document.

This patch removes importPaths() from the TextModifier.
The ViewerContext can be stored in the TextToModelMerger and
does not have to be recreated when needed.

Change-Id: I17281caee23ddd51f6e36d5346bc3bd7c53005e8
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@digia.com>