QmlDesigner: Adding a manual test for plugin integration
Because we had a couple of regressions in this area, I add this project
for manual testing.
In this project we have a qmlproject with a plugin that contains a .metainfo
file and a custom property sheet.

We define two items in the item library: MyComponent and MyComponent2.
While MyComponent has a custom property sheet the property sheet for
MyComponent2 is auto generated.

Change-Id: I8d5ade43f648354175587c0a3e989bad87149298
Reviewed-by: Tim Jenssen <tim.jenssen@digia.com>