KitModel: Fix crash on adding android qt version
createConfigWidget() calls KitManagerConfigWidget() calls discard(),
which results in a signal unmanagedKitUpdated. In response to that
signal, the KitModel iterates over all kit nodes. The node we are
currently creating doesn't have a widget yet, and thus leads to a crash.

Prevent that by adding the node only after the widget has been fully set
up.

Change-Id: I5b14be1c7363e480e3dd938853ff06c595993fc2
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>