C++: Show virtual functions menu only if resolving was requested
That is, if the user does Ctrl+LeftClick or (Ctrl+E,) F2, but not on
hover.

The described bug in the report occurred because we immediately jump to
an item if there is only one, e.g. in case of no overrides. That could
also happen on hover, which lead to the undesired effect.

Change-Id: I79f6fa1f90d37ea209da022d55d48aab2c13c21a
Task-number: QTCREATORBUG-12986
Reviewed-by: David Schulz <david.schulz@digia.com>