CppTools: Handle reset case for setIndexingSupport
For the time being, this function actually sets an additional
CppIndexSupport to be executed. Handle the case that a client tries to
restore the previous CppIndexingSupport (see SymbolSearcherTestCase).

This fixes GenericProjectManager::test_simple in interaction with
CppTools::test_builtinsymbolsearcher. Now first running the CppTools
tests and then the GenericProjectManager tests work fine.

Change-Id: Ibe7248ec8651d4fdc312f10cdcc9c56d6dd086d9
Reviewed-by: Christian Stenger <christian.stenger@digia.com>