CppTools: Fix C++ diagnostics hightlighting.
When the length of the highlight is specified, the code does not properly
check for end of line. This causes some lines to be highlighted when only
the last word was supposed to be.

For example, with this code:
   1: void foo(int x);
   2: int a = foo
   3:            ();
   4: int b = foo(
   5:             );
line 2 and 4 used to be completely highlighted (underlined), instead of
just 'foo'.

Change-Id: I40e895410ce0f38bad0adbccd509fd2943c93c97
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@theqtcompany.com>