Theming: Do not set the global palette
Setting the global palette is not a very good idea, because
it interferes with native styling.

On Windows 7 the dark theme was unuasable, because the text on
all natives controls was to bright.

All native controls and the global palette should be left alone.
Wherever we want styling/theming we have to use Manhatten style.

Change-Id: Ia1d0bfb190e4bde7be72f751afc0846570c1189d
Reviewed-by: hjk <hjk121@nokiamail.com>