Add CVS directory actions
This update introduces options to run "cvs update" and "cvs commit" on
a directory-by-directory basis in projects with multiple directories.
Two new entries are added to "CVS" menu: "Update Directory" and
"Commit Directory", which start the appropriate actions with the current
file's directory component as argument.

This might make it easier to control what exactly is synced with
the repository when working on a large project.

Task-number: QTCREATORBUG-13186
Change-Id: I620827211ca272403f1f0f99ebd8e61a096557d3
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>