Variable chooser: Handle geometry updates of (line)edit
The setTextMargin call which makes room for the button, can result in
sizeHint changes of the line edit, and therefore in relayouting and
geometry changes.

The patch also moves the escape key handling to the (line)edit instead
of unconditionally the focus widget, since there is code to hide the
chooser when the focus widget is not the (line)edit anyhow.

Change-Id: I687307985ca95e63c16086b8e77c371f642ab060
Task-number: QTCREATORBUG-13221
Reviewed-by: hjk <hjk121@nokiamail.com>
Reviewed-by: Christian Stenger <christian.stenger@digia.com>
Reviewed-by: Tim Jenssen <tim.jenssen@digia.com>