texteditor: add a hover handler which shows color preview tooltips
This adds a new hover handler which matches stand-alone color strings
like "#112233" or "Qt::yellow" or function argument tuples for colors
such as "QColor(0x11, 0x22, 0x33)". When matching against function
arguments, the function name must correspond to a recognized color
function (setRgb, etc. This is biased towards cpp text, but not limited
to such). The matching occurs when hovering over the arguments, not the
function.
If a match is identified, the hover handler gives it a relatively high
ranking.

Change-Id: Ied2927399cb19d6f562185a8b087f0ce118157db
Reviewed-by: David Schulz <david.schulz@theqtcompany.com>