SessionManager; configureEditors on project opening/parsing
So far editors were only configured on opening, which meant that if a
project was opened after the file was already open, the editor settings
were wrong. This commit fixes that by connecting to project opening.

Also due to changes in how qmake projects are parsed on initial opening
the project has a empty filelist. Connecting to fileListChanged() fixes
that too.

Task-number: QTCREATORBUG-13299
Change-Id: Ia648818c8c0adb9c6e5047b8c855b1f6790a7ae2
Reviewed-by: David Schulz <david.schulz@digia.com>