Git: Fix description of git status progress
Remove color.status option. Creator can handle ANSI now, so there
is no need to suppress that for user-visible output anymore (Git
is intelligent enough to not print colors, even if we do not explicitly
ask it not to:-)

For the internal uses of git status: Prefer using --porcelain instead
of the -s format. That is meant to be stable and does not use colors
ever.

All supported git versions (>= 1.7.2) support --porcelain.

Task-number: QTCREATORBUG-13277
Change-Id: Idd88bcba0c4e4ed68bb27fbf475fe88fae497f1c
Reviewed-by: Orgad Shaneh <orgads@gmail.com>