QbsPM: Do on-demand profile setup when profile name is requested.
This is necessary because the profile setup also entails putting the
profile name into the configuration we pass to the parsing step. If a
Kit has never been used to set up a qbs project, "qbs.profile" will not
exist in that configuration and the project cannot be parsed.

Task-number: QTCREATORBUG-13400
Change-Id: I32d767c36bd0ff33521137604a7826d9242a8712
Reviewed-by: Robert Loehning <robert.loehning@theqtcompany.com>
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>