Android: Don't create unsigned packages
The --release argument for androiddeployqt is really not very
useful, since it will create an unsigned package which cannot
be installed on a device or published in any way. This is a very
special use case, which I don't think we need to support in Creator,
and it should at least not be connected to the build flags for the
C++ code, since it's quite possible that you want to build your
application code in release, but still sign the apk with a debug key
to test it.

Task-number: QTCREATORBUG-13431
Change-Id: I752bfdf396eab596637ab111c7844128962db1fd
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>
Reviewed-by: BogDan Vatra <bogdan@kde.org>