JsonWizard: Add separate method to get stringified values
Have value(...) return a QVariant that is never mangeled and add
a stringValue(...) method to retrieve stringified values. It is
way easier to see who needs what now.

Use this consistently in the JsonWizard and fix one place where
a QVariant was expected but a stringified version was returned.

Task-number: QTCREATORBUG-13486
Change-Id: I2c4e9188280940e529f0f60bcc18b9b7330865d1
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>