Android: Fix deployment of release builds with Qt 5.3 and earlier
Since unsigned packages are not deployable, we always make
packages signed with the debug key when there's no signing requested,
even if the native code is built as release. This was already
fixed for Qt 5.4, where we use androiddeployqt to deploy the apk,
but for earlier versions, we need to use the correct apk name.
We also need to use the apk name which is used by androiddeployqt
in Qt 5.3, which does not contain the "-signed" suffix.

Change-Id: I87f7936002946acd80bc245935c44cf148f9630d
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>
Reviewed-by: BogDan Vatra <bogdan@kde.org>