ProjectExplorer: Fix disabled "Close Project" after closing project
The close project action is special. If there are multiple projects open
we show a menu with individual actions per project. If only one project
is open, we show one the "m_unloadAction". Even with only one project
open, the current project might be null. So explicitly manage enabled
the same way as visible. And also us the displayName of the startup
project, since that is guranteed to be non-null if there's a project
open.

Task-number: QTCREATORBUG-13497
Change-Id: I9150568da162c58ee524c2b954e0b9a816f79d7c
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>