SettingsDialog: Fix default button
The default button would move to the first pushbutton that got focus.
E.g. tabbing through the android settings would move the default button
to the first button that got focus.

The cause of that is that for QPushButton::setDefault to work, the
button needs to be a child of the QDialog. So call it after adding
the buttongroup to the layout.

Change-Id: Ife35bf15aa8585f7931a87ee715f316bb02b36be
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>