Doc: Improve description of "Qt mkspec" field
The mkspec is a global configuration for qmake, not a 'build instruction'.
For people new to qmake, and build systems in general, it might be arguably
a bit hard to grasp what an mkspec is (the qmake documentation doesn't
explain it apparently, either). Still, this is probably not the right
place to go much more into details ...

Change-Id: I80aa5fb9e4751242dd2570613951f2e1387c4433
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@theqtcompany.com>