Qt Quick Compiler Support: Only allow valid settings
If the Qt Quick Compiler binary is not present in the binary tools
directory of the Qt version, disable the checkbox. A warning will
be added in the next minor version, sind we already have string
freeze.

QML debugging does not work when using the Qt Quick Compiler. So,
in case the Qt Quick Compiler is checked, let's disable the QML
debugging checkbox.

Change-Id: I739b1f71dc53aca9cf00e6ffd15246184d89816b
Reviewed-by: hjk <hjk121@nokiamail.com>
Reviewed-by: Alessandro Portale <alessandro.portale@digia.com>