Core: Normalize given filePath in DocumentModel
TextDocument::open() sets a clean absolute path for the document, but
DocumentModel::documentForFilePath() did not normalize before comparing.

This resulted in the creation of multiple TextDocuments and Widgets for
the same path. Was re-producible with:

  1. Open the locator
  2. Type "f a.cpp" and Enter
  3. Go to 1.

Point 2 lead to an EditorManager::openEditor() call with an unclean
path. The second iteration triggered an QTC_ASSERT in CppModelManager.

Regression introduced by:

    commit 292cf6c9e87d2c293be6eb2808530dbbd728c8b6
    Algorithms for DocumentModel

Change-Id: I82fbc39251e48c1c97d6933857de172818d73f13
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>