QNX: Deploy only *.so.? libs
While on Unix systems, libQt5XYZ.so* are symlinks to libQt5XYZ.so.1.2.3, on
Windows hosts these files are copies instead of actual symlinks. Therefore,
when deploying from a Windows host, all the duplicates get uploaded to the
target, effectively taking a lot of unnecessary space.

Because Qt QNX apps link only agains *.so.? files, we deploy only those when
on Windows hosts.

Change-Id: I8f9f3f3a30d227b907f94f30ca47636dcb345eec
Reviewed-by: hjk <hjk121@nokiamail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>
Reviewed-by: Tobias Nätterlund <tobias.naetterlund@kdab.com>
Reviewed-by: Filipe Azevedo <filipe.azevedo@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>