CMake: Fix syntax highlighting for cmake files
The newest kate definition files for cmake contain no mimetype for the
cmake definitions, which lead to us not finding the right definition.

Instead add a method to find the definition based on the name, which
is not as generic, but we know which one we want for cmake anyway.

Task-number: QTCREATORBUG-13588
Change-Id: Ib57fe531a225310c32ca1ef909a31bd7f52c612f
Reviewed-by: Guilhem Vallat <guilhem.vallat@gmail.com>
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>
Reviewed-by: David Schulz <david.schulz@theqtcompany.com>