Fix "nmake ts-de" on Windows.
Local filenames should have the "file:///" prefix when passing them on to
the XQuery (if there is no host).

On Linux this worked without the prefix, while on windows QUrl got
confused by the colon (interpreted as the host:port separator).

For Unix filesystems, they will get their third slash from the first
slash in the absolute filename (when appended to the file:// prefix)

Change-Id: I41abed4821f1f253a13cc829eed8cb38e8e77bdc
Task-number: QTBUG-43417
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>