Coding Style: Clarify the use of the Q_OBJECT macro
The old definition sounded like strong but was qualified by an big
'if' at the end, resulting in stating the obvious for a single case
("use it when you need it for qobject_cast").

Having the reminder there is ok, if it is a common thing to forget,
but it should be recognizable as such, and can include "the other
case" where it is typically used.

Change-Id: I0183a12d14c414db9ff8b19598cb9dd22cedda8e
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@theqtcompany.com>
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>